# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-Powered Competitive Intelligence Platform",
    "description": "Market trend prediction and competitive intelligence API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/predictions": {
      "get": {
        "tags": [
          "Predictions"
        ],
        "summary": "List Predictions",
        "operationId": "list_predictions_api_v1_predictions_get",
        "parameters": [
          {
            "name": "company_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Company Id"
            }
          },
          {
            "name": "metric_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Metric Name"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PredictionResponse"
                  },
                  "title": "Response List Predictions Api V1 Predictions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predictions/models": {
      "get": {
        "tags": [
          "Predictions"
        ],
        "summary": "List Models",
        "operationId": "list_models_api_v1_predictions_models_get",
        "parameters": [
          {
            "name": "model_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Name"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModelVersionResponse"
                  },
                  "title": "Response List Models Api V1 Predictions Models Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predictions/models/{model_version_id}": {
      "get": {
        "tags": [
          "Predictions"
        ],
        "summary": "Get Model Version",
        "operationId": "get_model_version_api_v1_predictions_models__model_version_id__get",
        "parameters": [
          {
            "name": "model_version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Model Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelVersionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predictions/{company_id}": {
      "get": {
        "tags": [
          "Predictions"
        ],
        "summary": "Get Company Predictions",
        "operationId": "get_company_predictions_api_v1_predictions__company_id__get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Company Id"
            }
          },
          {
            "name": "metric_name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "market_activity_index",
              "title": "Metric Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/insights": {
      "get": {
        "tags": [
          "Insights"
        ],
        "summary": "List Insights",
        "operationId": "list_insights_api_v1_insights_get",
        "parameters": [
          {
            "name": "company_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Company Id"
            }
          },
          {
            "name": "insight_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Insight Type"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_InsightResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/insights/{insight_id}": {
      "get": {
        "tags": [
          "Insights"
        ],
        "summary": "Get Insight",
        "operationId": "get_insight_api_v1_insights__insight_id__get",
        "parameters": [
          {
            "name": "insight_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Insight Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/insights/generate": {
      "post": {
        "tags": [
          "Insights"
        ],
        "summary": "Generate Insight",
        "operationId": "generate_insight_api_v1_insights_generate_post",
        "parameters": [
          {
            "name": "company_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Company Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsightGenerateRequest",
                "default": {
                  "insight_types": [
                    "competitive",
                    "sentiment",
                    "trend",
                    "market_overview"
                  ],
                  "force": false
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ingestion/run": {
      "post": {
        "tags": [
          "Ingestion"
        ],
        "summary": "Trigger Ingestion",
        "description": "Trigger an ingestion run for a company.\n\nRuns in background so the API responds immediately.",
        "operationId": "trigger_ingestion_api_v1_ingestion_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestionTriggerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ingestion/runs": {
      "get": {
        "tags": [
          "Ingestion"
        ],
        "summary": "List Ingestion Runs",
        "operationId": "list_ingestion_runs_api_v1_ingestion_runs_get",
        "parameters": [
          {
            "name": "company_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Company Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_IngestionRunResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ingestion/runs/{run_id}": {
      "get": {
        "tags": [
          "Ingestion"
        ],
        "summary": "Get Ingestion Run",
        "operationId": "get_ingestion_run_api_v1_ingestion_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestionRunResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "company_id": {
            "type": "integer",
            "title": "Company Id"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "documents_found": {
            "type": "integer",
            "title": "Documents Found"
          },
          "documents_new": {
            "type": "integer",
            "title": "Documents New"
          },
          "documents_changed": {
            "type": "integer",
            "title": "Documents Changed"
          },
          "documents_skipped": {
            "type": "integer",
            "title": "Documents Skipped"
          },
          "requests_made": {
            "type": "integer",
            "title": "Requests Made"
          },
          "errors": {
            "type": "integer",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "started_at",
          "status",
          "documents_found",
          "documents_new",
          "documents_changed",
          "documents_skipped",
          "requests_made",
          "errors"
        ],
        "title": "IngestionRunResponse"
      },
      "IngestionTriggerRequest": {
        "properties": {
          "company_id": {
            "type": "integer",
            "title": "Company Id"
          },
          "source_types": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Types"
          }
        },
        "type": "object",
        "required": [
          "company_id"
        ],
        "title": "IngestionTriggerRequest"
      },
      "InsightGenerateRequest": {
        "properties": {
          "insight_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Insight Types",
            "description": "Types of insights to generate",
            "default": [
              "competitive",
              "sentiment",
              "trend",
              "market_overview"
            ]
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "description": "Force regeneration even if cached insight exists",
            "default": false
          }
        },
        "type": "object",
        "title": "InsightGenerateRequest"
      },
      "InsightResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "company_id": {
            "type": "integer",
            "title": "Company Id"
          },
          "insight_type": {
            "type": "string",
            "title": "Insight Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "evidence": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "insight_type",
          "title",
          "summary",
          "generated_at"
        ],
        "title": "InsightResponse"
      },
      "ModelVersionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "trained_at": {
            "type": "string",
            "format": "date-time",
            "title": "Trained At"
          },
          "training_samples": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Samples"
          },
          "features": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Features"
          },
          "mae": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mae"
          },
          "rmse": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rmse"
          },
          "mape": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mape"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "metadata_": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "model_name",
          "version",
          "trained_at",
          "status",
          "created_at"
        ],
        "title": "ModelVersionResponse"
      },
...
      "PaginatedResponse_IngestionRunResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IngestionRunResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "title": "PaginatedResponse[IngestionRunResponse]"
      },
      "PaginatedResponse_InsightResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InsightResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "title": "PaginatedResponse[InsightResponse]"
      },
...
      "PredictionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "company_id": {
            "type": "integer",
            "title": "Company Id"
          },
          "metric_name": {
            "type": "string",
            "title": "Metric Name"
          },
          "prediction_date": {
            "type": "string",
            "format": "date",
            "title": "Prediction Date"
          },
          "horizon_days": {
            "type": "integer",
            "title": "Horizon Days"
          },
          "predicted_value": {
            "type": "number",
            "title": "Predicted Value"
          },
          "lower_bound": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lower Bound"
          },
          "upper_bound": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upper Bound"
          },
          "model_version_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "metric_name",
          "prediction_date",
          "horizon_days",
          "predicted_value",
          "created_at"
        ],
        "title": "PredictionResponse"
      },
...
    }
  }
}